requirements.txt
omegaconf
matplotlib
numpy
torch
torchvision
torchmetrics
lightning
clearml
datasets
python-dotenv
tensorboard